requirements/runtime.txt
absl-py
accelerate>=0.19.0
cpm_kernels
datasets>=2.12.0
einops==0.5.0
evaluate>=0.3.0
func_timeout
fuzzywuzzy
h5py
immutabledict
jieba
json5
mmengine-lite
nltk==3.8
numpy>=1.23.4,<2.0.0
openai
OpenCC
opencv-python-headless
pandas<2.0.0
prettytable
protobuf
python-Levenshtein
rank_bm25==0.2.2
rapidfuzz
requests>=2.31.0
rich
rouge
-e git+https://github.com/Isaac-JL-Chen/rouge_chinese.git@master#egg=rouge_chinese
rouge_score
sacrebleu
scikit_learn==1.5.0
seaborn
sentence_transformers==2.2.2
tabulate
tiktoken
timeout_decorator
tokenizers>=0.13.3
torch>=1.13.1
tqdm>=4.64.1
transformers>=4.29.1
typer
